requirements.txt
Flask>=3.1.2
markdown2>=2.5.4
openai>=2.15.0
python-dotenv>=1.2.1
beautifulsoup4>=4.14.3
requests>=2.32.5
rich>=14.2.0
reportlab>=4.4.9
